requirements.txt
# Client
customtkinter>=5.2.0
Pillow>=10.0.0
pystray>=0.19.0
keyboard>=0.13.5
keyring>=25.0.0
requests>=2.31.0

# Server (деплоится отдельно на VPS)
# fastapi>=0.110.0
# uvicorn>=0.29.0
# sqlalchemy>=2.0.0
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.4

# Build
# pyinstaller>=6.0.0
